# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Employee Management System API",
        "description": "This is the API documentation for the Employee Management System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/employees/{employee_id}/attendance/departure": {
            "post": {
                "tags": [
                    "Attendance"
                ],
                "description": "Record an employee's departure",
                "operationId": "89411dc47104ef99a18e618e958073b2",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/EmployeeId"
                    }
                ],
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/RecordAttendanceSuccess"
                    },
                    "400": {
                        "$ref": "#/components/responses/DepartureAlreadyRecorded"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
...
        "/api/admins/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "description": "Register an admin",
                "operationId": "0d5d8f04a1a52ab8b17d75bade81b70f",
                "requestBody": {
                    "description": "Admin credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "$ref": "#/components/responses/RegisterValidationError"
                    },
                    "200": {
                        "$ref": "#/components/responses/RegisterSuccess"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/api/admins/reset-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "description": "Reset admin password",
                "operationId": "7c2a2172588d3b62636b605baa3b192a",
                "requestBody": {
                    "description": "Admin email, password and token",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ResetPasswordSuccess"
                    },
                    "403": {
                        "$ref": "#/components/responses/InvalidTokenOrExpiredError"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Admin": {
                "title": "Admin",
                "description": "Admin user schema",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "description": "The id of the admin",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the admin",
                        "type": "string"
                    },
                    "email": {
                        "description": "The email of the admin",
                        "type": "string",
                        "format": "email"
                    },
                    "created_at": {
                        "description": "The date and time the admin was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The date and time the admin was updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "example": {
                    "name": "Sheja Eddy",
                    "email": "admin@example.com",
                    "updated_at": "2024-10-10T12:00:00+00:00",
                    "created_at": "2024-10-10T12:00:00+00:00"
                }
            },
            "Attendance": {
                "title": "Attendance",
                "description": "Attendance model schema",
                "properties": {
                    "id": {
                        "description": "The id of the attendance",
                        "type": "integer"
                    },
                    "employee_id": {
                        "description": "The id of the employee",
                        "type": "integer"
                    },
                    "arrival_time": {
                        "description": "The arrival time of the employee",
                        "type": "string",
                        "format": "date-time"
                    },
                    "departure_time": {
                        "description": "The departure time of the employee",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "The date and time the attendance was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The date and time the attendance was updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "RegisterRequest": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "description": "The name of the admin",
                        "type": "string",
                        "example": "Sheja Eddy"
                    },
                    "email": {
                        "description": "The email of the admin",
                        "type": "string",
                        "format": "email",
                        "example": "admin@example.com"
                    },
                    "password": {
                        "description": "The password of the admin",
                        "type": "string",
                        "format": "password",
                        "example": "password"
                    },
                    "password_confirmation": {
                        "description": "Confirm password",
                        "type": "string",
                        "format": "password",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "ResetPasswordRequest": {
                "required": [
                    "email",
                    "password",
                    "token"
                ],
                "properties": {
                    "email": {
                        "description": "The email of the admin",
                        "type": "string",
                        "format": "email",
                        "example": "sheja@eddy.com"
                    },
                    "password": {
                        "description": "The new password of the admin",
                        "type": "string",
                        "format": "password",
                        "example": "password"
                    },
                    "password_confirmation": {
                        "description": "Confirm Password",
                        "type": "string",
                        "format": "password",
                        "example": "password"
                    },
                    "token": {
                        "description": "The reset token sent to the admin's email",
                        "type": "string",
                        "example": "d3b3c4d3-4b3d-4c3d-4b3d-4c3d4b3d4c3d"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "DepartureAlreadyRecorded": {
                "description": "If the departure for the employee is already recorded today or no arrival recorded for the employee",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "example": 400
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Departure already recorded for the employee today or no arrival recorded for the employee"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "InternalServerError": {
                "description": "Internal server error",
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "status",
                                "message"
                            ],
                            "properties": {
                                "status": {
                                    "description": "The status of the response",
                                    "type": "integer",
                                    "example": 500
                                },
                                "message": {
                                    "description": "The message of the response",
                                    "type": "string",
                                    "example": "Internal server error"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "InvalidTokenOrExpiredError": {
                "description": "Invalid token or token has expired",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "example": 403
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Invalid token or token has expired"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "RecordAttendanceSuccess": {
                "description": "Attendance recorded successfully",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "example": 201
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Attendance recorded successfully"
                                },
                                "data": {
                                    "$ref": "#/components/schemas/Attendance"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "RegisterSuccess": {
                "description": "Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "example": 200
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Success"
                                },
                                "data": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/Admin"
                                        },
                                        "token": {
                                            "description": "The token of the registered admin",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "RegisterValidationError": {
                "description": "Validation failed",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "example": 422
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Validation failed"
                                },
                                "errors": {
                                    "properties": {
                                        "name": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The name field is required"
                                            }
                                        },
                                        "email": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The email address is already in use"
                                            }
                                        },
                                        "password": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Passwords do not match"
                                            }
                                        },
                                        "password_confirmation": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The password confirmation field is required"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ResetPasswordSuccess": {
                "description": "Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "description": "The status of the response",
                                    "type": "integer",
                                    "example": 200
                                },
                                "message": {
                                    "description": "The message of the response",
                                    "type": "string",
                                    "example": "Password reset successfully"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "parameters": {
...
            "EmployeeId": {
                "name": "EmployeeId",
                "in": "path",
                "description": "Employee id parameter",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    }
}